Renamed Products to SpotOrders.  Previous commit also added RemoteConnection and DbConnectionConfig and factored out config path split.
