+++ src/FortitudeMarkets/Trading/ORX/Orders/Client/OrxOrderAmend.cs
@@ -7,7 +7,7 @@ using FortitudeCommon.DataStructures.Memory;
 using FortitudeCommon.Types.Mutable;
 using FortitudeIO.Protocols.ORX.Serdes;
 using FortitudeMarkets.Trading.Orders.Client;
-using FortitudeMarkets.Trading.Orders.Products;
+using FortitudeMarkets.Trading.Orders.SpotOrders;
 
 #endregion
