+++ src/FortitudeMarkets/Trading/ORX/Subscription/OrxTradingClient.cs
@@ -22,15 +22,15 @@ using FortitudeMarkets.Trading.Executions;
 using FortitudeMarkets.Trading.Orders;
 using FortitudeMarkets.Trading.Orders.Client;
 using FortitudeMarkets.Trading.Orders.Products;
-using FortitudeMarkets.Trading.Orders.Products.General;
 using FortitudeMarkets.Trading.Orders.Server;
+using FortitudeMarkets.Trading.Orders.SpotOrders;
 using FortitudeMarkets.Trading.Orders.Venues;
 using FortitudeMarkets.Trading.ORX.CounterParties;
 using FortitudeMarkets.Trading.ORX.Executions;
 using FortitudeMarkets.Trading.ORX.Orders;
 using FortitudeMarkets.Trading.ORX.Orders.Client;
-using FortitudeMarkets.Trading.ORX.Orders.Products.General;
 using FortitudeMarkets.Trading.ORX.Orders.Server;
+using FortitudeMarkets.Trading.ORX.Orders.SpotOrders;
 using FortitudeMarkets.Trading.ORX.Orders.Venues;
 using FortitudeMarkets.Trading.ORX.Session;
 
@@ -261,6 +261,7 @@ public class OrxTradingClient : OrxHistoricalTradesClient, ITradingFeedListener
                     var orxOrder = order switch
                                    {
                                        ISpotOrder spotOrder => new OrxSpotOrder(spotOrder)
+                                     , _ => new OrxSpotOrder((ISpotOrder)order)
                                    };
