+++ src/FortitudeMarkets/Trading/Orders/Client/IOrderAmend.cs
@@ -2,6 +2,7 @@
 
 using FortitudeCommon.DataStructures.Memory;
 using FortitudeMarkets.Trading.Orders.Products;
+using FortitudeMarkets.Trading.Orders.SpotOrders;
 
 #endregion
