+++ src/FortitudeTests/FortitudeMarkets/Trading/ORX/Orders/Products/General/OrxSpotOrderTests.cs
@@ -9,7 +9,8 @@ using FortitudeIO.Protocols.ORX.Serdes.Serialization;
 using FortitudeIO.Protocols.Serdes.Binary.Sockets;
 using FortitudeMarkets.Trading.Orders.Products;
 using FortitudeMarkets.Trading;
-using FortitudeMarkets.Trading.ORX.Orders.Products.General;
+using FortitudeMarkets.Trading.Orders.SpotOrders;
+using FortitudeMarkets.Trading.ORX.Orders.SpotOrders;
 using FortitudeMarkets.Trading.ORX.Orders.Venues;
 
 #endregion
